// database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
// ...
      opportunities: {
        Row: {
          company: string
          created_at: string | null
          deadline: string | null
          id: string
          is_rolling: boolean | null
          link: string | null
          location: string | null
          logo_emoji: string | null
          pay_hr: number | null
          pay_monthly: number | null
          tags: string[] | null
          title: string
          type: string | null
        }
        Insert: {
          company: string
          created_at?: string | null
          deadline?: string | null
          id?: string
          is_rolling?: boolean | null
          link?: string | null
          location?: string | null
          logo_emoji?: string | null
          pay_hr?: number | null
          pay_monthly?: number | null
          tags?: string[] | null
          title: string
          type?: string | null
        }
        Update: {
          company?: string
          created_at?: string | null
          deadline?: string | null
          id?: string
          is_rolling?: boolean | null
          link?: string | null
          location?: string | null
          logo_emoji?: string | null
          pay_hr?: number | null
          pay_monthly?: number | null
          tags?: string[] | null
          title?: string
          type?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}